requirements.txt
openslide-python==1.1.1
scipy==1.1.0
imageio==2.4.1
imagesize==1.1.0
scikit-image==0.14.0